backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq==0.13.0
manim==0.17.3
numpy==1.24.3
psutil==5.9.6
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
watchdog==2.3.1
starlette==0.27.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.027.0
